feat: added new feature for handling checkbox state and scrollable section for our disclaimer.
Implemented a new feature to correctly update the checkbox state in the AthleteRegistration component. User is required to agree to terms before going to next page. Also Implemented a scrollable section so the disclaimer is not taking up the entire bottom of the page.

fix: ensure event object is passed to handleAthleteSubmit

Updated the button click event in AthleteRegistration.jsx to pass the event object to handleAthleteSubmit, preventing the 'TypeError: Cannot read properties of undefined (reading 'preventDefault')' error.